requirements.txt
nltk==3.9.1
tensorflow==2.17.0
tf-keras==2.17.0
transformers==4.45.2
flask==3.0.3
Flask-Limiter==3.8.0
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
gunicorn==20.1.0

# test requirements
pytest-flask==1.3.0

# dev requirements
black==24.10.0